move testCommon to test/util
